feat: Test AdapterViews with Espresso
In this exercise you'll finish building the MenuActivityScreenTest,
which demos a user clicking on a GridView item in MenuActivity,
and opens up the corresponding OrderActivity.

(51, 4) // TODO (1) Add annotation to specify AndroidJUnitRunner class as
the default test runner
(55, 8) // TODO (2) Add the rule that provides functional testing of a
single activity
(60, 8) // TODO (3) Finish writing this test which will click on a gridView
Tea item and verify that the OrderActivity opens up with the correct
tea name displayed.
(27, 8) // TODO (4) Add the required Espresso UI Testing dependencies