feat: Add IdlingResourceActivityTest in TeaTime
In this exercise you will implement the IdlingResourceActivityTest. We will
create a mock scenario in the app that simulates a time delay when downloading
menu images and descriptions into the MenuActivity.

To help you get started I’ve added the SimpleIdlingResource class which is an
implementation of IdlingResource for you to use.

You will need to modify MenuActivity to properly implement the interface
ImageDownloader which is a callback interface. Refer back to the previous
concepts if you need a hint.

When the callback finishes executing, it will return the arraylist of Tea
objects with the images and tea names back to the MenuActivity.

==============================================================================

(35, 4) // TODO (1) Implement ImageDownloader.DelayerCallback
(43, 8) // TODO (2) Add a SimpleIdlingResource variable that will be null in
production
(47, 8) * TODO (3) Create a method that returns the IdlingResource variable.
It will instantiate a new instance of SimpleIdlingResource if the
IdlingResource is null. This method will only be called from test.
(63, 8) * TODO (4) Using the method you created, get the IdlingResource
variable. Then call downloadImage from ImageDownloader. To ensure there's
enough time for IdlingResource to be initialized, remember to call
downloadImage in either onStart or onResume. This is because @Before in
Espresso Tests is executed after the activity is created in onCreate, so there
might not be enough time to register the IdlingResource if the download is
done too early.
(118, 8) // TODO (5) Override onDone so when the thread in ImageDownloader
is finished, it returns an ArrayList of Tea objects via the callback.
(71, 8) // TODO (6) Registers any resource that needs to be synchronized with
Espresso before the test is run.
(81, 8) // TODO (7) Test that the gridView with Tea objects appears and we can
click a gridView item
(87, 8) // TODO (8) Unregister resources when not needed to avoid malfunction